requirements.txt
asgiref==3.5.0
Django==4.0.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
